from /Users/yohei/Desktop/src_101007.tar.gz
git-svn-id: http://istr.iis.u-tokyo.ac.jp/svn/istr/trunk@10 575e276f-71f8-4965-9d1f-d4eeffdffdfb

+++ src/AssyMatrix.cpp
@@ -78,7 +78,7 @@ CAssyMatrix::CAssyMatrix(/* const */ CAssyModel *pAssyModel)
 			};
 		};
 		mvMPCMatrix.push_back(mpc);
-		printf("mvMPCMatrix.size() %d \n", mvMPCMatrix.size());
+		printf("mvMPCMatrix.size() %d \n", (uint)mvMPCMatrix.size());
 	};
 }
